Extracted AppLayoutMenu class
